move getAnimators from simple adapter to UltimateViewAdapter
